+++ test/support/test_invoice_subscriber.exs
@@ -1,0 +1,3 @@
+defmodule TestInvoiceSubscriber do
+  use LndClient.InvoiceUpdatesSubscriber
+end